Change Codeception to output in XML
The parallel processing of unit tests can only merge XML result files. Since
the Codeception tasks in FutureCI make use of JSON instead of XML I had to
change either the JSON to XML in the Codeception Parallel actions (which
would be time consuming) or change FutureCI to use XML as output instead
of JSON.

The latter proved trivial so in this commit we have replaced the JSON output
with XML and matched the tests to verify this.