Use the suite name in the config if provided
In this commit I have altered the algorith for retrieving the Build CN
(name used to make the build path with) to use the build.name variable
in the configuration if it exists.

If the name does not exist it will fallback to the current behaviour
and use the filename.

The Build CN is not a one-to-one match with the name; any uppercase letters
are converted to lowercase and spaces are converted to underscores to match
the existing format.