Add log-dir option to Exec task
The exec does not copy its artifacts from a target folder to the build
artifacts folder. This causes an issue with the CodeceptionResultParser
that expects the 'report.xml' to be in the build artifacts folder.

To fix this issue I have added a(n optional) option called log-dir
in the Exec task that copies anything in the given folder to the
build artifacts folder so that the CodeceptionResultParser will be
able to the find the reports.xml file when invoking Codeception
using the exec task (as we do for the functional tests).