Go back to build listing after viewing a build
During testing I discovered that it was not possible to return to the
home page and build listing after viewing a build's details.

After some research this was caused because the Object.prototype was
extended; something which jQuery does not accept.

Because the build details page was the first to use this feature it
was also the first one to show problematic behaviour.

To fix this I have removed all code regarding the extension of
Object.prototype and changed all parts of the code that makes use
of these extensions.