Add Formatter infrastructure
In this commit I have added the basics needed to support Formatters in FutureCI.
This feature will allow the user to perform actions after the build has completed and
the result object has been populated. These actions can be used to generate additional
artifacts that can then be used in the view for FutureCI or can be offered as download.

See https://github.com/f500/future-ci/issues/3 for more information