Add commit information to output
Before this commit FutureCI was not aware of any commit information
during the build.

With this commit I have added a Git Subscriber to the Run Command
that retrieves the commit information from the root folder in the
configuration and sets a Commit entity object on the current Build.

With this information any other subscriber, such as the Slack and
Console subscriber are able to read and display additional information
about the commit that initiated this build.